# services/api/gen/api.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for QuickAddParsedType.
const (
	QuickAddParsedTypeExpense QuickAddParsedType = "expense"
	QuickAddParsedTypeIncome  QuickAddParsedType = "income"
)

// Valid indicates whether the value is a known member of the QuickAddParsedType enum.
func (e QuickAddParsedType) Valid() bool {
	switch e {
	case QuickAddParsedTypeExpense:
		return true
	case QuickAddParsedTypeIncome:
		return true
	default:
		return false
	}
}

// Defines values for TransactionType.
const (
	TransactionTypeExpense TransactionType = "expense"
	TransactionTypeIncome  TransactionType = "income"
)

// Valid indicates whether the value is a known member of the TransactionType enum.
func (e TransactionType) Valid() bool {
	switch e {
	case TransactionTypeExpense:
		return true
	case TransactionTypeIncome:
		return true
	default:
		return false
	}
}

// Defines values for TransactionCreateType.
const (
	Expense TransactionCreateType = "expense"
	Income  TransactionCreateType = "income"
)

// Valid indicates whether the value is a known member of the TransactionCreateType enum.
func (e TransactionCreateType) Valid() bool {
	switch e {
	case Expense:
		return true
	case Income:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	Code    string    `json:"code"`
	Details *[]string `json:"details,omitempty"`
	Message string    `json:"message"`
}

// FamilyGroup defines model for FamilyGroup.
type FamilyGroup struct {
	CreatedAt *time.Time          `json:"created_at,omitempty"`
	CreatedBy *openapi_types.UUID `json:"created_by,omitempty"`
	Id        openapi_types.UUID  `json:"id"`
	Name      string              `json:"name"`
}

// FamilyGroupCreate defines model for FamilyGroupCreate.
type FamilyGroupCreate struct {
	Name string `json:"name"`
}

// MonthlyReport defines model for MonthlyReport.
type MonthlyReport struct {
	Balance      float32             `json:"balance"`
	Categories   *map[string]float32 `json:"categories,omitempty"`
	TotalExpense float32             `json:"total_expense"`
	TotalIncome  float32             `json:"total_income"`
}

// QuickAddParsed defines model for QuickAddParsed.
type QuickAddParsed struct {
	Amount      *float32            `json:"amount,omitempty"`
	Description *string             `json:"description,omitempty"`
	Type        *QuickAddParsedType `json:"type,omitempty"`
}

// QuickAddParsedType defines model for QuickAddParsed.Type.
type QuickAddParsedType string

// Transaction defines model for Transaction.
type Transaction struct {
	Amount          float32             `json:"amount"`
	CreatedAt       *time.Time          `json:"created_at,omitempty"`
	Currency        string              `json:"currency"`
	Date            openapi_types.Date  `json:"date"`
	Description     *string             `json:"description,omitempty"`
	GroupId         *openapi_types.UUID `json:"group_id,omitempty"`
	Id              openapi_types.UUID  `json:"id"`
	RecurringRuleId *openapi_types.UUID `json:"recurring_rule_id,omitempty"`
	Type            TransactionType     `json:"type"`
	UserId          openapi_types.UUID  `json:"user_id"`
}

// TransactionType defines model for Transaction.Type.
type TransactionType string

// TransactionCreate defines model for TransactionCreate.
type TransactionCreate struct {
	Amount      float32               `json:"amount"`
	Currency    string                `json:"currency"`
	Date        openapi_types.Date    `json:"date"`
	Description *string               `json:"description,omitempty"`
	GroupId     *openapi_types.UUID   `json:"group_id,omitempty"`
	Type        TransactionCreateType `json:"type"`
}

// TransactionCreateType defines model for TransactionCreate.Type.
type TransactionCreateType string

// TransactionList defines model for TransactionList.
type TransactionList struct {
	Data []Transaction `json:"data"`

	// NextCursor Cursor for the next page of results. Null if no more pages.
	NextCursor *string `json:"next_cursor"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// GetMonthlyReportParams defines parameters for GetMonthlyReport.
type GetMonthlyReportParams struct {
	Year    int                 `form:"year" json:"year"`
	Month   int                 `form:"month" json:"month"`
	GroupId *openapi_types.UUID `form:"group_id,omitempty" json:"group_id,omitempty"`
}

// ListTransactionsParams defines parameters for ListTransactions.
type ListTransactionsParams struct {
	// GroupId Filter by family group
	GroupId *openapi_types.UUID `form:"group_id,omitempty" json:"group_id,omitempty"`

	// Cursor Cursor for pagination (e.g., encoded date + id)
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`

	// Limit Number of items to return
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// QuickAddTransactionJSONBody defines parameters for QuickAddTransaction.
type QuickAddTransactionJSONBody struct {
	// Text Syntactic string generated by speech to text voice API.
	Text string `json:"text"`
}

// CreateGroupJSONRequestBody defines body for CreateGroup for application/json ContentType.
type CreateGroupJSONRequestBody = FamilyGroupCreate

// CreateTransactionJSONRequestBody defines body for CreateTransaction for application/json ContentType.
type CreateTransactionJSONRequestBody = TransactionCreate

// QuickAddTransactionJSONRequestBody defines body for QuickAddTransaction for application/json ContentType.
type QuickAddTransactionJSONRequestBody QuickAddTransactionJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List family groups
	// (GET /groups)
	ListGroups(w http.ResponseWriter, r *http.Request)
	// Create a family group
	// (POST /groups)
	CreateGroup(w http.ResponseWriter, r *http.Request)
	// Get monthly summary report
	// (GET /reports/monthly)
	GetMonthlyReport(w http.ResponseWriter, r *http.Request, params GetMonthlyReportParams)
	// List transactions
	// (GET /transactions)
	ListTransactions(w http.ResponseWriter, r *http.Request, params ListTransactionsParams)
	// Create a transaction
	// (POST /transactions)
	CreateTransaction(w http.ResponseWriter, r *http.Request)
	// Parse voice transcribed text to a transaction
	// (POST /transactions/quick-add)
	QuickAddTransaction(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// List family groups
// (GET /groups)
func (_ Unimplemented) ListGroups(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a family group
// (POST /groups)
func (_ Unimplemented) CreateGroup(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get monthly summary report
// (GET /reports/monthly)
func (_ Unimplemented) GetMonthlyReport(w http.ResponseWriter, r *http.Request, params GetMonthlyReportParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List transactions
// (GET /transactions)
func (_ Unimplemented) ListTransactions(w http.ResponseWriter, r *http.Request, params ListTransactionsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a transaction
// (POST /transactions)
func (_ Unimplemented) CreateTransaction(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Parse voice transcribed text to a transaction
// (POST /transactions/quick-add)
func (_ Unimplemented) QuickAddTransaction(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListGroups operation middleware
func (siw *ServerInterfaceWrapper) ListGroups(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListGroups(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateGroup operation middleware
func (siw *ServerInterfaceWrapper) CreateGroup(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateGroup(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetMonthlyReport operation middleware
func (siw *ServerInterfaceWrapper) GetMonthlyReport(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetMonthlyReportParams

	// ------------- Required query parameter "year" -------------

	if paramValue := r.URL.Query().Get("year"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "year"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "year", r.URL.Query(), &params.Year, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "year", Err: err})
		return
	}

	// ------------- Required query parameter "month" -------------

	if paramValue := r.URL.Query().Get("month"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "month"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "month", r.URL.Query(), &params.Month, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "month", Err: err})
		return
	}

	// ------------- Optional query parameter "group_id" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "group_id", r.URL.Query(), &params.GroupId, runtime.BindQueryParameterOptions{Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "group_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMonthlyReport(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListTransactions operation middleware
func (siw *ServerInterfaceWrapper) ListTransactions(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListTransactionsParams

	// ------------- Optional query parameter "group_id" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "group_id", r.URL.Query(), &params.GroupId, runtime.BindQueryParameterOptions{Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "group_id", Err: err})
		return
	}

	// ------------- Optional query parameter "cursor" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "cursor", r.URL.Query(), &params.Cursor, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "cursor", Err: err})
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListTransactions(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateTransaction operation middleware
func (siw *ServerInterfaceWrapper) CreateTransaction(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateTransaction(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// QuickAddTransaction operation middleware
func (siw *ServerInterfaceWrapper) QuickAddTransaction(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.QuickAddTransaction(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/groups", wrapper.ListGroups)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/groups", wrapper.CreateGroup)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/reports/monthly", wrapper.GetMonthlyReport)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/transactions", wrapper.ListTransactions)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/transactions", wrapper.CreateTransaction)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/transactions/quick-add", wrapper.QuickAddTransaction)
	})

	return r
}